# src/main/resources/openapi/api.yaml
openapi: 3.1.0
info:
  title: Tracker - OpenAPI 3.1
  description: description
  version: 1.0.0
servers:
  - url: http://localhost/api/v1
tags:
  - name: product
    description: Operations about products
paths:
  /product:
    post:
      tags:
        - product
      summary: Add a new product
      description: Add a new product
      operationId: addProduct
      requestBody:
        description: Create a new product
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
        required: true
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid input
    get:
      tags:
        - product
      summary: Finds products
      description: Finds products
      operationId: findProducts
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
  /product/{productId}:
    get:
      tags:
        - product
      summary: Find product by ID
      description: Returns a single product
      operationId: getProductById
      parameters:
        - name: productId
          in: path
          description: ID of product to return
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Invalid ID supplied
        '404':
          description: Product not found
    delete:
      tags:
        - product
      summary: Deletes a product
      description: delete a product
      operationId: deleteProduct
      parameters:
        - name: productId
          in: path
          description: Product id to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '400':
          description: Invalid product value
components:
  schemas:
    Product:
      required:
        - name
      type: object
      properties:
        id:
          type: string
          format: uuid
        description:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double